Fixed an issue where skill sprite coords where output with "width" and "height" instead of "w" and "h"
