+++ PoESkillTree.Engine.GameModel/PassiveTree/Base/JsonPassiveTreeSkillSprite.cs
@@ -10,16 +10,16 @@ namespace PoESkillTree.Engine.GameModel.PassiveTree.Base
         [JsonProperty("filename")]
         public string FileName { get; set; } = string.Empty;
 
-        [JsonProperty("coords", ItemConverterType = typeof(PassiveTreeRectangleFConverter))]
+        [JsonProperty("coords", ItemConverterType = typeof(PassiveTreeSkillSpriteCoordsConverter))]
         public Dictionary<string, RectangleF> Coords { get; set; } = new Dictionary<string, RectangleF>();
     }
 
     public class JsonPassiveTreeOldSkillSprite : JsonPassiveTreeSkillSprite
     {
-        [JsonProperty("notableCoords", ItemConverterType = typeof(PassiveTreeRectangleFConverter))]
+        [JsonProperty("notableCoords", ItemConverterType = typeof(PassiveTreeSkillSpriteCoordsConverter))]
         public Dictionary<string, RectangleF> NotableCoords { get; private set; } = new Dictionary<string, RectangleF>();
 
-        [JsonProperty("keystoneCoords", ItemConverterType = typeof(PassiveTreeRectangleFConverter))]
+        [JsonProperty("keystoneCoords", ItemConverterType = typeof(PassiveTreeSkillSpriteCoordsConverter))]
         public Dictionary<string, RectangleF> KeystoneCoords { get; private set; } = new Dictionary<string, RectangleF>();
     }
 }
